+++ VMDeploy.Management/changelog.md
@@ -1,5 +1,12 @@
 ﻿# Changelog
 
-## 1.0.0 (???)
+## ???
+
++ Upd: Configuration: Template - added property "DynamicHardwareProfile"
++ Upd: Configuration: Template - added property "DynamicGuestOSProfile"
++ Upd: JEA Endpoint - added new commands from VMDeploy.Orchestrator
++ Fix: Import-VMManConfiguration - creates action folder if not present yet
+
+## 1.0.0 (2022-05-16)
 
 + Initial Release